connected front-end and back-end after restructuring and installing + running necessary dependencies and scripts
