+++ client/package.json
@@ -1,27 +1,21 @@
 {
-  "name": "react-app",
+  "name": "client",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://localhost:3001/",
   "dependencies": {
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.1",
-    "bcrypt": "^5.0.1",
-    "bootstrap": "^4.6.0",
-    "connect-session-sequelize": "^7.1.1",
-    "express": "^4.17.1",
-    "express-session": "^1.17.2",
-    "mongoose": "^5.12.14",
-    "mysql2": "^2.2.5",
+    "bootstrap": "^5.0.2",
     "react": "^17.0.2",
     "react-bootstrap": "^1.6.1",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
     "react-router-dom": "^5.2.0",
     "react-script-tag": "^1.1.2",
-    "react-scripts": "4.0.3",
-    "sequelize": "^6.6.2",
+    "react-scripts": "^4.0.3",
     "three": "^0.129.0",
     "web-vitals": "^1.1.2"
   },
