+++ client/src/components/Asset.js
