+++ controllers/api/coinRoutes.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const { User, Coin } = require('../../models');
-const withAuth = require('../../scripts/auth');
+const withAuth = require('../../utils/auth');
 
 router.get('/', async (req, res) => {
   try {
