+++ controllers/api/index.js
@@ -7,4 +7,5 @@ console.log("THIS IS INDEX.JS");
 router.use('/users', userRoutes);
 router.use('/coins', coinRoutes);
 
+
 module.exports = router;
