+++ controllers/homeRoutes.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const { Coin, User } = require('../models');
-const withAuth = require('../scripts/auth');
+const withAuth = require('../utils/auth');
 
 router.get('/', async (req, res) => {
   try {
