--- react-app/.env
@@ -1,5 +1,0 @@
-API_KEY=7b425e31-3d77-4f19-81b4-62d933a34f22
-
-DB_NAME=coin_db
-DB_PASSWORD=g1j2g1j2
-DB_USER=root
