--- react-app/README.md
@@ -1,2 +1,0 @@
-# project-3
-Project3
