--- react-app/routes/api/api.js
@@ -1,21 +1,0 @@
-const rp = require('request-promise');
-const requestOptions = {
-  method: 'GET',
-  uri: 'https://pro-api.coinmarketcap.com/v1/cryptocurrency/listings/latest',
-  qs: {
-    'start': '1',
-    'limit': '20',
-    'convert': 'USD'
-  },
-  headers: {
-    'X-CMC_PRO_API_KEY': '7b425e31-3d77-4f19-81b4-62d933a34f22'
-  },
-  json: true,
-  gzip: true
-};
-
-rp(requestOptions).then(response => {
-  console.log('API call response:', response);
-}).catch((err) => {
-  console.log('API call error:', err.message);
-});
