+++ seeds/coinData.json
