+++ server.js
@@ -3,6 +3,8 @@ const express = require('express');
 const session = require('express-session');
 const routes = require('./controllers');
 
+require('dotenv').config();
+
 const sequelize = require('./config/connection');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
